requirements.txt
zenml[server]==0.55.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pydantic==2.5.3
pyyaml==6.0.1

